+++ packages/engine/CHANGELOG.md
@@ -3,7 +3,7 @@
 ## v1.1.0
 
 - FIX: Added the needed styles to the body tag to avoid issues with the Flutter app.
-- Added new attestationToken option
+- Added new options/params to the engine: attestationToken, onAuthTokenError and onRequestAddFunds
 - Saved singleton instances in the global scope to be accessible by the Flutter app using JS interoperability: https://dart.dev/interop/js-interop
 
 ## v1.0.0